tech(TTB-060): Moved Header component to individual directory
